Allow the phone home stats to be able to run on any worker.
